Satisfy Windows compiler

+++ crypto_kem/hqc-192/clean/reed_solomon.c
@@ -35,10 +35,10 @@ void PQCLEAN_HQC192_CLEAN_reed_solomon_encode(uint8_t *cdw, const uint8_t *msg) 
         }
 
         for (size_t k = PARAM_N1 - PARAM_K - 1; k; --k) {
-            cdw[k] = cdw[k - 1] ^ tmp[k];
+            cdw[k] = (uint8_t)(cdw[k - 1] ^ tmp[k]);
         }
 
-        cdw[0] = tmp[0];
+        cdw[0] = (uint8_t)tmp[0];
     }
 
     memcpy(cdw + PARAM_N1 - PARAM_K, msg, PARAM_K);
